Major overhaul to the way that DAQ and PDF information is accumulated to speed up likelihood evaluation.
When generating a likelihood, the DAQ can be run many times in parallel
by the GPU, creating a large block of channel hit information in memory.
The PDF accumulator processes that entire block in two passes:

  * First update the channel hit count, and the count of channel hits
    falling into the bin around the channel hit time being evaluated.
    Add any channel hits that should also be included in the n-th
    nearest neighbor calculation to a channel-specific work queue.

  * Process all the work queues for each channel and update the
    list of nearest neighbors.

This is hugely faster than what we were doing before.  Kernel
estimation (or some kind of orthogonal function expansion of the PDF)
should be better ultimately, but for now the nearest neighbor approach
to PDF estimation seems to be working the best.