Add a summing mode to the event view.
Press "s" and average time, charge and hit probability are computed
for all of the events in the input file.  Then you can cycle through
the different values using the "." key, just as you can for single
events.  You can flip between sum and regular modes, and the sum
calculation is only done the first time.