Apply a floor to the probability of each channel being in the observed state (hit or not), rather than a floor on the hit probability.
A channel that is impossible to hit should have zero probability in
the likelihood and not be hit in the actual data.  Before this change,
that channel would be forced to have a hit probability of 0.5 /
ntotal, which is wrong.  We only need to ensure the probability of the
observed state of the channel is not zero so that the log() function
is defined.