Split photon propagation in the likelihood calculation into a "forced scatter" and a "forced no-scatter" pass.
Since we want to include contributions from two populations of weighted
photons, we have to break up the DAQ simulation into three functions:

 * begin_acquire()
 * acquire()
 * end_acquire()

The first function resets the channel states, the second function
accumulates photoelectrons (and can be called multiple times), and the
last function returns the hit information.  A global weight has also
been added to the DAQ simulation if a particular set of weighted
photons need to have an overall penalty.

The forced scattering pass can be repeated many times on the same
photons (with the photons individually deweighted to compensate).
This reduces the variance on the final likelihoods quite a bit.