Cache class for managing the geometry cache on disk.
The storage format is changing relative to the old format, so all
geometry files will be saved in the ~/.chroma/geo directory.  For now,
the format is just a simple pickle.  We know this is not optimal for
space or speed, but the Geometry class will be changing soon, and we
can optimize further after that.

This Cache class will also soon manage the separate BVH cache.