Speed up node intersection by 2.5x using tips from "Optimizing ray tracing for CUDA" by Hannu Saransaari.
The intersect_box() function has been rewritten to be much shorter and
use the min() and max() functions, which map directly to hardware
instructions.  Additionally, the calculations inside intersect_box()
have been reorganized to allow the compiler to use the combined
multiply-add instruction, instead of doing a subtraction followed by a
division (which is way slower).