Skip L1 cache when loading nodes.
Node access is very irregular as each thread descends the BVH tree.
Each node is only 16 bytes, so the 128 byte cache line size in the L1
cache means that a lot of useless data is often fetched.  Using some
embedded PTX, we can force the L1 cache to be skipped, going directly
to L2.  The L2 cache line is 32 bytes long, which means that both
children in a binary tree will be cached at the same time.

This improves the speed on the default generated binary trees, but
does not help an optimized tree yet.